Create password authentication
Project 2 : Python Project

1.	Send Automatic Emails using Python
Content – 
Every time you register on a new app, you automatically receive a welcome message with your name on it. If you want to learn how to send such emails automatically, then this article is for you.

 You will automatically send a welcome message to the newly registered user. For this task, you must first generate a google app password for your Gmail account. Once you’ve generated your Google app password,  you can start the task of sending emails using Python by writing the logic for code.
Password Authentication using Python
Content :
Password Authentication is the process of checking the identity of a user. Almost every online platform today makes sure that they only give access to the real user which can be only possible by asking for a password while a user wants to log in to the account
To create a password authentication system using Python you have to follow the steps mentioned below:
1.	Create a dictionary of usernames with their passwords.
2.	Then you have to ask for user input as the username by using the input function in Python.
3.	Then you have to use the getpass module in Python to ask for user input as the password. Here we are using the getpass module instead of the input function to make sure that the user doesn’t get to see what he/she write in the password field.
![image](https://github.com/user-attachments/assets/51b3c5bf-3643-4d99-9303-dd2a037f98f7)